[tlscommon] Add test helper methods
